+++ src/plugins/simple/aside.ts
@@ -2,12 +2,13 @@ import Layout from "../../enum/layout";
 import Global from "../../util/global";
 import Util from "../../util/util";
 
-function renderSidebarItem(list: Array<any>,init: boolean){
+function renderSidebarItem(list: Array<any>,init: boolean) {
+    let active = Util.getActiveMenu(list);
     let html = "";
     if(init){
         html = `
             <el-menu
-                default-active="1"
+                default-active="${active}"
                 class="el-menu-vertical-demo"
                 :default-openeds="menuOpens">
         `;
@@ -16,10 +17,10 @@ function renderSidebarItem(list: Array<any>,init: boolean){
         if(!item.children){
             html += `<el-menu-item index="${item.index}"><a class="nav-a" href="${item.href}" target="${item.target}">${item.title}</a></el-menu-item>`
             if(item.href.startsWith("#")){
-                Util.setNavMap(item.href,item.title);
+                Util.setNavMap(item.href, item.title);
             }
         }else {
-            let itemHtml = renderSidebarItem(item.children,false);
+            let itemHtml = renderSidebarItem(item.children, false);
             html += `
                 <el-submenu index="${item.index}">
                     <template slot="title">
@@ -45,7 +46,7 @@ export default {
         .then(response => response.text())
         .then(json => {
             let sidebarList = JSON.parse(json);
-            let html = renderSidebarItem(sidebarList,true);
+            let html = renderSidebarItem(sidebarList, true);
             // console.info("sidebar html:"+html);
             elSide.innerHTML = html;
             let pageEl: any = document.getElementById(Layout.main);